implement add note functionality
the AddNote button now takes user to a form and allows them to add a new note w/ title, content, and they select an existing folder to put it in. validation is performed and then the folderId is added to state so that the note is associated with the folder in the rest of the app. the only thing that remains is to put a timestamp on the note object.